Changes in light of version notation change

+++ spec/spec_helper.rb
@@ -28,9 +28,11 @@ RSpec.configure do |c|
   c.treat_symbols_as_metadata_keys_with_true_values = true
 
   c.before(:suite) do
-    @rest = Diametric::RestService.new("spec/test_version_file.yml", "tmp/datomic")
-    @rest.start(:port => 46291, :db_alias => @storage, :uri => "datomic:mem://")
-    PID = @rest.pid
+    unless :service || :transactor
+      @rest = Diametric::RestService.new("spec/test_version_file.yml", "tmp/datomic")
+      @rest.start(:port => 46291, :db_alias => @storage, :uri => "datomic:mem://")
+      PID = @rest.pid
+    end
     if ENV['TRANSACTOR']
       FileUtils.cp(File.join('spec', 'config', 'logback.xml'), File.join('bin', 'logback.xml'))
     end
@@ -38,7 +40,9 @@ RSpec.configure do |c|
 
   c.after(:suite) do
     Diametric::Persistence::Peer.shutdown(true) if is_jruby?
-    Process.kill("HUP", PID)
+    unless :service || :transactor
+      Process.kill("HUP", PID)
+    end
     if ENV['TRANSACTOR']
       FileUtils.rm(File.join('bin', 'logback.xml'), :force => true)
     end
